fix(scope): $emit and $broadcast take array arg instead of varargs
Breaking Change!

before: scope.$emit(eventName, message1, message2):
after: scope.$emit(eventName, [message1, message2]);

before: scope.$broadcast(eventName, message1, message2);
after: scope.$broadcast(eventName, [message1, message2]);

The reasoning for this change is that varargs make the api
non-extensible in the future. So for example when we add cancelable
broadcast events, we'll need to be able to control the cancelability
somehow and for this we'll need an additional argument.