feat($route): allow async resolution of route depedencies
!!! EXPERIMENTAL CHANGE FOR DFA!!!

This change adds init property to the route definition.

The value of the init property is a injectable function which returns a
map, which will be resolved as map of promises and the resulting map
of the promise values will then be used as locals during the injection
of the controller specified for this route.

If a single promise in the map fails, the route change is aborted and if
the rejection reason is in form {redirectTo: '/somePath'} then a
redirection to this path will occur.

Current drawbacks
- lots of ugly code in route definition
- the init function is moderately hard to test
- doesn't prefetch the template along with all the other dependencies
- doesn't safely handle situation when a subsequent route change is
  fired off while the first one is still waiting for all promises to be
  resolved.